+++ aiproteomics/modelgen/test_prosit_to_onnx.py
@@ -1,5 +1,5 @@
 from onnx import load
-from prosit1_model import build_prosit1_model
+from prosit1_gen import build_prosit1_model
 import os
 
 def shape2tuple(shape):
